backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
GeoAlchemy2==0.15.2
geopandas==1.0.1
gotrue==2.8.1
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
postgrest==0.16.11
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pyogrio==0.10.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==2.0.2
scipy==1.13.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
websockets==12.0
Werkzeug==3.0.3
yarl==1.12.1
